QBasic online: fix wrong mouse position
This is an ugly hack. The program is that the DOS cursor position is not
being tracked until the first click. After your first click, it starts
following the movement of the cursor but will be offset incorrectly.

(The movement X and movement Y values take priority, not the
absolute position values)

Fix: After the first click, the next time the mouse moves, we change its
X and Y movement value to be as if it had moved from 0,0 to the current
position in that frame.